3D solid export
Combine base and NMR mesh to produce a 3D solid that is recognised by CAD software.

Note: Fusion 360 gives error messages that the mesh is not closed and that it does not have a positive volume, however it still allows toolpaths to be generated.

Note: Sharp peaks will not print very well. It would be a good idea to add some smoothing.